Sync & bugs fix
-Text update
-EndSpeaker
-knight rework
-ring of dominate
-gun with ring of  accuracy
-magestaff

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/KnightSKILL.java
@@ -280,7 +280,7 @@ public class KnightSKILL extends Buff implements ActionIndicator.Action {
                     if (counter.count() < 10) {
                         counter.countUp(1);
                         if (counter.count() == 10) GLog.p(Messages.get(KnightSKILL.class, "combo_max"));
-                        else GLog.i(Messages.get(KnightSKILL.class, "combo_stack", counter.count()));
+                        else GLog.i(Messages.get(KnightSKILL.class, "combo_stack", (int)counter.count()));
                     }
                     break;
                 case SMASH:
@@ -288,7 +288,7 @@ public class KnightSKILL extends Buff implements ActionIndicator.Action {
                     if (counter.count() < 10) {
                         counter.countUp(1);
                         if (counter.count() == 10) GLog.p(Messages.get(KnightSKILL.class, "combo_max"));
-                        else GLog.i(Messages.get(KnightSKILL.class, "combo_stack", counter.count()));
+                        else GLog.i(Messages.get(KnightSKILL.class, "combo_stack", (int)counter.count()));
                     }
                     break;
                 case KILLBLOW:
@@ -296,7 +296,7 @@ public class KnightSKILL extends Buff implements ActionIndicator.Action {
                     if (counter.count() < 10) {
                         counter.countUp(1);
                         if (counter.count() == 10) GLog.p(Messages.get(KnightSKILL.class, "combo_max"));
-                        else GLog.i(Messages.get(KnightSKILL.class, "combo_stack", counter.count()));
+                        else GLog.i(Messages.get(KnightSKILL.class, "combo_stack", (int)counter.count()));
                     }
                     break;
                 case LIGHTSWORD:
